# jara-market-backend/jaraman/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jaramarket API",
        "description": "Jaramarket Meals Marketplace API for Customer and Vendor mobile applications",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "L5_SWAGGER_CONST_HOST",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Order History",
                "description": "Retrieve a list of orders placed by the authenticated customer.",
                "operationId": "064eb18304b7962f884ad58571453d98",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Place Order",
                "description": "Create a new order with multiple products or ingredients.",
                "operationId": "33d737c93193573b1687e463b1562642",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_date",
                                    "delivery_type",
                                    "service_charge",
                                    "total"
                                ],
                                "properties": {
                                    "order_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-05-20"
                                    },
                                    "delivery_type": {
                                        "type": "string",
                                        "example": "delivery",
                                        "enum": [
                                            "pickup",
                                            "delivery"
                                        ]
                                    },
                                    "shipping_fee": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    },
                                    "service_charge": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "vat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5650
                                    },
                                    "remarks": {
                                        "type": "string",
                                        "example": "Please deliver between 2pm and 4pm."
                                    },
                                    "address_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 2500
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "ingredients": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "ingredient_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1500
                                                },
                                                "unit": {
                                                    "type": "string",
                                                    "example": "kg"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{order}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Order Details",
                "description": "Retrieve detailed information about a specific order.",
                "operationId": "bc66a3e4b8c0fc651e9f55737f5207e0",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "The Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{order}/cancel": {
            "put": {
                "tags": [
                    "Customer"
                ],
                "summary": "Cancel Order",
                "description": "Cancel a pending order.",
                "operationId": "b4389cd103363c6a60ef3be71116908c",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "The Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}